backend/requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
distro==1.8.0
dnspython==2.4.0
docx2pdf==0.1.8
exceptiongroup==1.1.2
fastapi==0.109.1
frozenlist==1.4.0
google-cloud-storage==2.10.0
h11==0.14.0
httpcore>=1.*
httptools==0.6.0
httpx==0.27.2
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
langchain==0.3.9
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.55.3
ollama==0.4.2
orjson==3.9.15
pdf2image==1.16.3
protobuf>=4.24.4
pydantic==2.10.2
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.7
pytz==2023.3
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.0
scipy>=1.5.2
transformers==4.36.0
tokenizers>=0.14
tqdm==4.65.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.23.1
uvloop==0.20.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
pillow>=10.0.0 # not directly required, pinned by Snyk to avoid a vulnerability
loguru>=0.7.2
pinecone-client>=3.1.0
sentence-transformers>=2.5.1
beautifulsoup4
lxml>=5.2.0
lightrag-hku==1.0.1
aioboto3==13.2.0
tiktoken==0.8.0 
nano-vectordb==0.0.4.3
neo4j==5.27.0
oracledb==2.5.0
debugpy==1.8.0  